.config.yaml: Fix kas warnings
When using refspec instead of commit/tag/branch, kas produces a
warning for each repository:

```
WARNING  - Using deprecated refspec for repository ${repository}
You should migrate to commit/tag/branch.
```

While "branch" is the same as refspec, it is not desirable to use,
because kas will pull the tip of the branch whenever building from
scratch, which could result in bugs or other build problems.

As such, we should use tags first, as they are specific commits in a
human-readable format, and then specific commits if a tag is not
available.

Change-Id: Iefd45e9fc90636cc23fb5f5b4913647871f56433
Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>