voyager.conf: add opengl, pam, vulkan, and wayland
These are features that are required to build the gstreamer packages and
weston package. If the feature is not used, then the above packages are
not built. As such, it is safe to add these features to voyager.conf,
even if there is an image without weston, as no additional space is
used.

Change-Id: I727b87d1f1538b939d7ccfa9259a95018596af5e
Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>