meta-axelera: recipes-core: Improve container integration
It is not optimal to have the recipe itself download the container image
and integrate it into the rootfs, since it takes a very long time to
build and flash the board each time.
Also, the steps required to get the container working are somehow
complicated and cumbersome.

Fix both of these things by adding a setup script which handles
everything for us:
- Starts docker service
- Downloads the docker image
- Loads the image into docker
- Copies the SDK from the container
- Changes ownership of the SDK directory

In this way, using the container is simple and efficient.

Change-Id: I923b9c0733f1c87c04f8c12472f8da946dcd5d3a
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>