meta-axelera: import parted v3.6 from openembedded-core
Parted v3.6 is required for our use case, because we need to automate
the `fix` user prompt when launching parted as a script, and the `-f`
option that does this is not present in v3.4, which is the one included
in our version of `openembedded-core`.

Import v3.6 from `openembedded-core` at the link below.

Link: https://layers.openembedded.org/layerindex/recipe/397760/
Change-Id: I0deeafe9ccfa7706acf23135c6994ce52fd29e60
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>