axelera-container: Remove manual docker service start
The docker service is working as intended: the docker.socker
automatically starts and is listening, while the docker.service with the
daemon itself is ment to start once the first docker command is invoked.

Remove this part of the script, since it is redundant and it allows to
make this script necessary only on first boot.

The issue of the very long docker starting time (about 3 minutes)
is unrelated and still needs to be investigated.

Change-Id: Ied139575d163b42fc8ca83a7373df0455effb8d3
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>