meta-axelera: docker-ce: Fix loading time issue
On each reboot, the first time any docker-related
command is called or the docker service is manually started it takes
more than 120 seconds.

After thorough investigation, this was traced down to the 120 second
default timeout in:
`systemd-networkd-wait-online.service`

To avoid this, one possibility was to change the timeout value to a
smaller one (or 0).

Since we do not require the network to be strictly online,
simply work around the issue by moving the dependency from
`network-online.target` to `network.target`.

As this a workaround, it still requires further investigation into
the cause of the problem itself, which is that the network
interfaces timeout.

Change-Id: I4de7157015e7597d53f2c2943d14a2ba28f01523
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>