kernel-modules: Update the driver for arm recovery
This commit updates the recovery thread timings to a number which
makes Omega recovery work. The old Alpha timings end up not
detecting a link down during reset because they are too big.

This Related:SDK-5119 in the sense that it will help testing things
there but doesn't yet completely solve the problem.
It just makes it so that if there's a fail:
the next run or networks doesn't need any fixes on the host: the
recovery mechanism will just make the next run work

Change-Id: I64ba39da57b0232baa4137ddc3813a11430ce035
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>