axelera-container: Fix setup script
Following the changes on userid and groupid that map `ubuntu` group
inside the container with `axelera` group on the host, the ownership
change of the setup script was raising some errors.

Since the cause for the ownership change has been removed by mapping the
groupids, fix this by simply removing it.

Change-Id: I5da9c61ae187f7a5b8cb4f3d3e5c395555ede8e5
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>