kernel-module-axelera: Fix directory structure
At current state, there are two similarly named and redundant
directories.
Fix this by moving the single patch that is being applied to the other
similar directory. With this move, the patch file is also being renamed
to correct a typo in the word `compille`.

Additionally, since auto-patch is not needed anymore, remove it from the
recipe inheritance.

Change-Id: I697f45790b2572a655e9e9d6a69e4adc6105b4cd
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>