kernel-modules: Fix check pcie device script
At current state, the script uses a tracker file placed in a volatile
directory. This means that if the PCIe device is not found, then it
touches a file in a volatile directory, and reboots. Because of this at
the first instance of a missing PCIe card, the board is stuck in an
infinite reboot loop.
Fix this by changing the directory of the tracker file from `/var/log`
to `/var/local`.

Change-Id: Ida970d6c6f2b54f297212bf1ac2eb0138011f7b4
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>