voyager-image: Change home directory groups
Change home directory groups from each respective user to the common
axelera, and change permissions from 0755 to 0775 to allow other users
in the axelera group to write in all home directories.

Do this because it allows weston user to write into them, which means
that when using the weston GUI one does not have to have root permission
to have the same workflow.

Change-Id: Ibd376ed40af0c576f95624bda57f5e109095057e
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>