axelera-container: Change setup script
Following permission changes, the start_axelera python script can create
the voyager-sdk directory within the user home directories.
Thanks to this, that part of the setup script is not necessary anymore,
since this can be directly handled by start_axelera.

Additionally, the permission change for weston access is now also
handled inside start_axelera.

Remove the docker instructions and the permission change.
Keep this script for internal Amarula comodity, to allow us to handle
the download and SHA256 checksum of the archive and the docker load.

YouTrack-Id: AXELERA-57
Change-Id: Ie89e001a691503efaa9af77adf3b6bc25098e435
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>