axelera-container: Move display permission change to start_axelera
Since we want to add the possibility of painting on the display from
inside the container, adding a check to the start_axelera script is
fine for our use-case.

Add a check if a display is connected; if yes, change its weston
permissions such that the container will be allowed to access it.

YouTrack-Id: AXELERA-57
Change-Id: Id604294e20698dccb178f940b0d2401d7541f7ba
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>