axelera-container: Add persistent directory
Following explicit request, add a persistent directory to share data
between the container and the host machine.

The directory itself is called `shared`, and is located in the home
directory of the `antelao` or `firefly` users on the host machine, and
in the home directory of the `ubuntu` user within the container.

YouTrack-Id: AXELERA-65
Change-Id: I293e6c4260f97ac946414e24c852ee3a07066032
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>