voyager-image-weston: Add xfce4-terminal
The requirement is for the terminal to support 24-bit true color, in
order to correctly display the output of some software suite regarding
the AI accelerator card.

The current terminal emulator, weston-terminal, does not support this
feature.

To fix this, add xfce4-terminal to the Voyager weston image, with all of
its dependencies:
- Modified distro features, as x11 and xwayland are requirements for
  xfce4-terminal.
- Modified weston.ini, such that the icon shown in the top left to open
  the terminal emulator now points to xfce4-terminal instead of
  weston-terminal.

YouTrack-Id: AXELERA-67
Change-Id: I044b20ec13f9dafc98c69801b09edfffdbbe01a2
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>