xfce4-terminal: Change default encoding to UTF-8
Even after adding a terminal emulator supporting full 24-bit color, the
output of Axelera inference script was not displaying correctly.
The issue was because of the default encoding for the xfce4-terminal,
which is ANSI_X3.4-1968 (ASCII).

Fix this by adding a bbappend recipe which adds a configuration file for
xfce4-terminal, explicitly setting the default encoding to UTF-8.

YouTrack-Id: AXELERA-67
Change-Id: Ib687f413abd3064dfc14eb9a042501560390806d
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>