Add support for the testusb tool
It is required to provide some objective tests about USB functionality
for the Antelao board.

Because of this, add support for the `testusb` tool.
The tool itself is a part of the kernel source, and is designed to test
USB ports in various use-cases to provide objective pass/fail results.

YouTrack-Id: AXELERA-83
Change-Id: I042e2946f09a9fd1767221ea8d3c45e850efd250
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>