setup_axelera_enviroment: Do not hardcode the SDK version number
Allow the user to download any version of the Axelera SDK container.
The user specifies the version as a command-line argument to
setup_axelera_environment.sh. All downloaded containers and associated
checksum files will be stored in a local directory named after the
corresponding version. For example, `bash setup_axelera_environment.sh
1.3.0-rc2` will download the container archive and the corresponding
checksum file (if found on S3) to directory `v.1.3.0-rc2`. If a
checksum file is not found (either locally or remotely), the integrity
check will be skipped.

Change-Id: I365ecd851665ea9ad8915703fc458a511e0b1061
Signed-off-by: Ioannis Koltsidas <ioannis.koltsidas@axelera.ai>
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>