meta-axelera: Add custom os-release
It was explicitly requested to us that we add a way to recognize which
device the BSP is running on to os-release.
Moreover, it is useful for client support purposes that we include a way
to know exactly which build is running on the device, especially ones
coming from the Jenkins CI.

Because of this, add a bbappend recipe for os-release.
The new /etc/os-release will be of the following form:

	ID=voyager
	NAME="Voyager Linux"
	VERSION="1.1.0-rc9"
	BUILD_ID="<jenkins_build_id>"
	PRETTY_NAME="Voyager Linux 1.1.0-rc9"
	BOARD_TYPE=antelao-3588

Note that the new custom field `BUILD_ID` defaults to the internal
`DATETIME` BitBake variable (e.g. "YYYYMMDDhhmmss") if
`JENKINS_BUILD_ID` is not injected to kas-container with:
	--runtime-args "-e JENKINS_BUILD_ID=<jenkins_build_id>"

YouTrack-Id: AXELERA-95
Change-Id: I2627c4fd50d78af2fbf47cdceb09eea6690912f0
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>