kas/include/base.yaml: Add JENKINS_BUILD_ID to env
We plan to inject this variable to kas-container from the Jenkins CI,
to have a consistent way of identifying which specific Jenkins build is
running on-device.

This will be useful when receiving client or customer support requests,
to be able to easily reach the same build and try to reproduce and fix
any issues.

YouTrack-Id: AXELERA-95
Change-Id: I82bbd69e9d0ae17ae79a128795372a1263b1611a
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>