axelera.wss.config: Remove non-existant `.tgz` from Mend check
When mend checking the project, a number of packages (and therefore
vulnerabilities) were being included even if not present.
This issue was traced back to the fact that Mend, by default, uses
package managers to search for all dependencies of products and adds all
of them in the report, even if not present in the source.

Fix this by explicitly telling mend to not do this, by setting
`resolveAllDependencies` to false.

Change-Id: Iaadc267036265db860065f1644262c07238fbaf0
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>