Update functionality in start_axelera.py to match usage
The script will check if the requested version exists as image and it's
mandatory to provide `--container-name` and `--version`. The help has
been improved and printed messages are not allign. The user is reminded
to run `make clobber-libs && make operators` once the container starts
and this is one time job.

Change-Id: If41a6022520239312a13003fc860710c9066a035