setup-directories: new recipe
This installs a .service file and script to /usr/bin/ that runs on startup.
The job of this script is to set up any directories that may be missing
or needed on startup. It's useful for a read-only file system.

This will be extended in the future to create directories for mender and
overlay mount points.

YouTrack-Id: AXELERA-80
Change-Id: Icc34b61de317c063fceba10c730ec0d18fff2c4d
Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>