Add u3loop suite for Passmark USB3.0 Loopback Plug
Passmark only provide a proprietary software for communicating with
the device, which only works on Windows.
For Linux compatibility, they suggest using BurnInTest, which is
another closed and paid software.

Add recipe for u3loop.
This is a open-source utility (MIT License) written by David Imhoff
used to communicate with the Passmark USB3.0 Loopback Plug on Linux.

YouTrack-Id: AXELERA-100
Change-Id: I8fd33e08428ab83946433b98f667b4484cd29268
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>