mender-commit-check: new recipe
Add a recipe that installs a shell script that auto-checks to see
if the /data/mender/scripts directory exists. This directory only
exists when an update has been installed and a reboot has happened
but the update has not yet been commited.

If the directory exists, the script runs `mender commit`.
If /data/mender/modules exists, those are also removed as the modules
directory is also provided by update artifacts and is not used otherwise.

YouTrack-Id: AXELERA-80
Change-Id: I727a0a1c22e433f80a4125185530b6e5c4a0f732
Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>