auto-extend-partition: Remove `parted` command
After moving from having a single rootfs partition to three different
partitions (/, data, factory) on the eMMC, the `parted` command has been
failing to extend mmcblk0p5, giving an error because it is currently
mounted.

This feature was working correctly when attempting to on-line resize the
rootfs on mmcblk0p3.

Since practically all users are flashing the board with the Rockchip
.update.img through `upgrade_tool UF`, which automatically extends the
last partition to fill the eMMC already, simply remove the `parted`
command and only leave the `resize2fs`.

Note that this patch effectively breaks proper usage when flashing the
`.wic` image with `upgrade_tool WL 0`.

Change-Id: Ia20e88a6ab2a2511f763df376a68d43b9ecff296
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>