overlayfs-etc-preinit.sh.in: Handle rw edge case
- Rockchip U-Boot temporarily loads the rootfs as read-write.
- All of the the auto-generated services by `volatile-binds` contain:
	ConditionPathIsReadWrite=!/<path>

There is an edge case in which these volatile-binds services are ran
BEFORE the rootfs is remounted as read-only, resulting in the
above-mentioned condition NOT being met.
This makes it so that the overlays are not mounted, making the system
unuseable.

Handle this by forcing the rootfs to be remounted as read-only in the
preinit script.

Change-Id: I4410c0a67ac945298054360313b571dc09e20e3b
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>