linux-rockchip: Support package manager
There are some issues with the package manager on the linux-based system
caused by the init script having a non-standard name.
Since the package manager is required for our use-case to update the
metis driver without updating the system itself on the antelao board,
we need to add the package manager feature on the BSP.

To fix the issue we must add the preinit script to the firefly and
antelao bootargs in both their dts.

Since the original `chosen` of the firefly board is inherited by the
`rk3588-linux.dtsi`, copy it to make the change less important for other
boards and to make it easier to maintain.
The newly added chosen is exactly the same as the original, with
`init=/sbin/preinit` appended at the end.

Change-Id: I47c0d71c245197f8541e6a8572bbe752cb8d1711
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>