mender-client: Add dependencies to growfs service
Right now, mmcblk0p6 is being correctly extended by Rockchip
upgrade_tool, but the `df -h` output still shows:

Filesystem                Size      Used Available Use% Mounted on

/dev/mmcblk0p6          114.5M     89.0K    105.4M   0% /data
/data/overlay/etc/upper 114.5M     89.0K    105.4M   0% /etc
overlay                 114.5M     89.0K    105.4M   0% /home

The cause of this is that `mender-systemd-growfs-data.service` is
failing:

x mender-systemd-growfs-data.service - Grow File System on /data
     Loaded: loaded
        (/lib/systemd/system/mender-systemd-growfs-data.service; static)
     Active: failed (Result: exit-code) since
                      Thu 1970-01-01 00:00:03 UTC; 55 years 7 months ago
    Process: 316 ExecStart=/lib/systemd/systemd-growfs /data
                                         (code=exited, status=1/FAILURE)
   Main PID: 316 (code=exited, status=1/FAILURE)

The most likely cause of this is that the service is being executed too
early in the chain.
To fix this, add a custom `mender-systemd-growfs-data.service` which
adds After and Requires of the `systemd-udev-settle.service`, forcing it
to wait for all pending udev events to be handled before being executed.

YouTrack-Id: AXELERA-80
Change-Id: I2e2ec2c7c137c579a91484db7235235e19281557
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>