meta-voyager: Add cve-extra-exclusions
This file contains a list of CVE's where resolution has proven to be
impractical or there is no reasonable action the Yocto Project can take
to resolve the issue.
It contains all the information we are aware of about an issue and
analysis about why we believe it can't be fixed/handled. Additional
information is welcome through patches to the file.

Change-Id: I15af7506bb0e928676a2dad11f3239f7dd147917
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>