os-release.bbappend: add GIT_HASH
Add the git hash of the meta-axelera repository
to /etc/os-release. If a git hash doesn't exist, then the
GIT_HASH field is set to NONE.

Also, add `do_compile[nostamp] = "1`. Without this set, the
git hash is cached and not re-read during every build.
Even with this change, the build remains reproducible if the
git hash has not changed.

Forcing a recompile during every build incerases the build
time by approximately 10 seconds.

YouTrack-Id: AXELERA-116
Change-Id: I23537932acb1b9270652540c4e559ef79f28d3c8
Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>