meta-axelera: Add post leave script to flash the boot.img in boot partition
At current state using mender to update the board while changing the dts
causes it to break and get stuck in U-Boot.

Fix this issue by manually copying the boot.img from the new partition,
to respect the Rockchip-specific boot process and allowing us to update
the dts as well using Mender.

YouTrack-Id: AXELERA-138
Change-Id: I8e8b2bb75d742391193823e57817692f4dc9375d
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>