meta-axelera: Move the uboot enviroment to proper partition
Avoid to have the enviroment in the middle of partition but
allocate them on a fixed partition and fixed position. To ensure
it create fixed-size partition for all the partition defined
and adjust mender and script with the new layout. This patch
add the flashing of the env1 and env2 and generate the binary
file used to do it.

YouTrack-Id: AXELERA-138
Change-Id: I051d923cde6e652a758a447bcf50cfcd190bc13d
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>