meta-axelera: Fix partitions for itx-3588j
It was discovered that the firefly board was having trouble with the
uboot environment.
This was traced down to a faulty partition scheme which was only working
for antelao.

Fix this by tweaking the uboot environment, the mender config and the
wks file to allow both boards to work as intended.

YouTrack-Id: AXELERA-145
Change-Id: Ie1ffaf9c2807b0fa733c13794554764cc22234e8
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>